Move microbit rendering code to separate file, and make sure that createLessonPdf waits until microbit blocks have been rendered. (#19)
